# Fix `function-no-unknown` performance by reducing file read count