--- a/.changeset/quick-gifts-shop.md
+++ b/.changeset/quick-gifts-shop.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `function-no-unknown` performance by reducing file read count