--- a/lib/rules/function-no-unknown/index.cjs
+++ b/lib/rules/function-no-unknown/index.cjs
@@ -25,6 +25,9 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/function-no-unknown',
 };
 
+/** @type {string[] | undefined} */
+let functionsListMemo;
+
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions) => {
 	return (root, result) => {
@@ -45,7 +48,10 @@ const rule = (primary, secondaryOptions) => {
 			return;
 		}
 
-		const functionsList = JSON.parse(fs.readFileSync(functionsListPath.toString(), 'utf8'));
+		// Read the file lazily for performance.
+		functionsListMemo ??= JSON.parse(fs.readFileSync(functionsListPath.toString(), 'utf8'));
+
+		const functionsList = /** @type {string[]} */ (functionsListMemo);
 
 		root.walkDecls((decl) => {
 			const { value } = decl;