--- a/lib/rules/function-no-unknown/index.cjs
+++ b/lib/rules/function-no-unknown/index.cjs
@@ -2,9 +2,8 @@
 // please instead edit the ESM counterpart and rebuild with Rollup (npm run build).
 'use strict';
 
-const fs = require('node:fs');
+const node_module = require('node:module');
 const cssParserAlgorithms = require('@csstools/css-parser-algorithms');
-const functionsListPath = require('css-functions-list');
 const cssTokenizer = require('@csstools/css-tokenizer');
 const validateTypes = require('../../utils/validateTypes.cjs');
 const declarationValueIndex = require('../../utils/declarationValueIndex.cjs');
@@ -15,6 +14,13 @@ const report = require('../../utils/report.cjs');
 const ruleMessages = require('../../utils/ruleMessages.cjs');
 const validateOptions = require('../../utils/validateOptions.cjs');
 
+var _documentCurrentScript = typeof document !== 'undefined' ? document.currentScript : null;
+const require$1 = node_module.createRequire((typeof document === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : (_documentCurrentScript && _documentCurrentScript.src || new URL('lib/rules/function-no-unknown/index.cjs', document.baseURI).href)));
+
+// TODO: Replace this with Import attributes when dropping the support for older Node.js versions.
+// See https://nodejs.org/api/esm.html#import-attributes
+const functionsList = require$1('css-functions-list/index.json');
+
 const ruleName = 'function-no-unknown';
 
 const messages = ruleMessages(ruleName, {
@@ -24,9 +30,6 @@ const messages = ruleMessages(ruleName, {
 const meta = {
 	url: 'https://stylelint.io/user-guide/rules/function-no-unknown',
 };
-
-/** @type {string[] | undefined} */
-let functionsListMemo;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions) => {
@@ -47,11 +50,6 @@ const rule = (primary, secondaryOptions) => {
 		if (!validOptions) {
 			return;
 		}
-
-		// Read the file lazily for performance.
-		functionsListMemo ??= JSON.parse(fs.readFileSync(functionsListPath.toString(), 'utf8'));
-
-		const functionsList = /** @type {string[]} */ (functionsListMemo);
 
 		root.walkDecls((decl) => {
 			const { value } = decl;