--- a/lib/rules/function-no-unknown/index.mjs
+++ b/lib/rules/function-no-unknown/index.mjs
@@ -23,6 +23,9 @@ const meta = {
 	url: 'https://stylelint.io/user-guide/rules/function-no-unknown',
 };
 
+/** @type {string[] | undefined} */
+let functionsListMemo;
+
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions) => {
 	return (root, result) => {
@@ -43,7 +46,10 @@ const rule = (primary, secondaryOptions) => {
 			return;
 		}
 
-		const functionsList = JSON.parse(fs.readFileSync(functionsListPath.toString(), 'utf8'));
+		// Read the file lazily for performance.
+		functionsListMemo ??= JSON.parse(fs.readFileSync(functionsListPath.toString(), 'utf8'));
+
+		const functionsList = /** @type {string[]} */ (functionsListMemo);
 
 		root.walkDecls((decl) => {
 			const { value } = decl;