--- a/lib/rules/function-no-unknown/index.mjs
+++ b/lib/rules/function-no-unknown/index.mjs
@@ -1,7 +1,6 @@
-import fs from 'node:fs';
+import { createRequire } from 'node:module';
 
 import { isFunctionNode, parseListOfComponentValues, walk } from '@csstools/css-parser-algorithms';
-import functionsListPath from 'css-functions-list';
 import { tokenize } from '@csstools/css-tokenizer';
 
 import { isRegExp, isString } from '../../utils/validateTypes.mjs';
@@ -13,6 +12,12 @@ import report from '../../utils/report.mjs';
 import ruleMessages from '../../utils/ruleMessages.mjs';
 import validateOptions from '../../utils/validateOptions.mjs';
 
+const require = createRequire(import.meta.url);
+
+// TODO: Replace this with Import attributes when dropping the support for older Node.js versions.
+// See https://nodejs.org/api/esm.html#import-attributes
+const functionsList = require('css-functions-list/index.json');
+
 const ruleName = 'function-no-unknown';
 
 const messages = ruleMessages(ruleName, {
@@ -22,9 +27,6 @@ const messages = ruleMessages(ruleName, {
 const meta = {
 	url: 'https://stylelint.io/user-guide/rules/function-no-unknown',
 };
-
-/** @type {string[] | undefined} */
-let functionsListMemo;
 
 /** @type {import('stylelint').Rule} */
 const rule = (primary, secondaryOptions) => {
@@ -45,11 +47,6 @@ const rule = (primary, secondaryOptions) => {
 		if (!validOptions) {
 			return;
 		}
-
-		// Read the file lazily for performance.
-		functionsListMemo ??= JSON.parse(fs.readFileSync(functionsListPath.toString(), 'utf8'));
-
-		const functionsList = /** @type {string[]} */ (functionsListMemo);
 
 		root.walkDecls((decl) => {
 			const { value } = decl;