--- a/lib/utils/mathMLTags.cjs
+++ b/lib/utils/mathMLTags.cjs
@@ -7,12 +7,11 @@ const node_module = require('node:module');
 var _documentCurrentScript = typeof document !== 'undefined' ? document.currentScript : null;
 const require$1 = node_module.createRequire((typeof document === 'undefined' ? require('u' + 'rl').pathToFileURL(__filename).href : (_documentCurrentScript && _documentCurrentScript.src || new URL('lib/utils/mathMLTags.cjs', document.baseURI).href)));
 
-// NOTE: mathml-tag-names v3 is a pure ESM package,
+// TODO: mathml-tag-names v3 is a pure ESM package,
 // so we cannot update it while supporting both ESM and CommonJS.
 //
 // In addition, mathml-tag-names v2 provides only a JSON file,
 // so ESM cannot import it (raises the "ERR_IMPORT_ASSERTION_TYPE_MISSING" error).
-// @ts-expect-error -- mathml-tag-names v2 doesn't have type declarations.
-const mathMLTags = require$1('mathml-tag-names');
+const mathMLTags = require$1('mathml-tag-names/index.json');
 
 module.exports = mathMLTags;