--- a/lib/utils/mathMLTags.cjs
+++ b/lib/utils/mathMLTags.cjs
@@ -12,6 +12,7 @@ const require$1 = node_module.createRequire((typeof document === 'undefined' ? r
 //
 // In addition, mathml-tag-names v2 provides only a JSON file,
 // so ESM cannot import it (raises the "ERR_IMPORT_ASSERTION_TYPE_MISSING" error).
+/** @type {string[]} */
 const mathMLTags = require$1('mathml-tag-names/index.json');
 
 module.exports = mathMLTags;