--- a/lib/utils/mathMLTags.mjs
+++ b/lib/utils/mathMLTags.mjs
@@ -7,6 +7,7 @@ const require = createRequire(import.meta.url);
 //
 // In addition, mathml-tag-names v2 provides only a JSON file,
 // so ESM cannot import it (raises the "ERR_IMPORT_ASSERTION_TYPE_MISSING" error).
+/** @type {string[]} */
 const mathMLTags = require('mathml-tag-names/index.json');
 
 export default mathMLTags;