<!-- Each pull request must be associated with an open issue unless it's a documentation fix. If a corresponding issue does not exist, please create one so we can discuss the change first. -->

<!-- Please answer the following. We close pull requests that don't. -->

> Which issue, if any, is this issue related to?

None.

> Is there anything in the PR that needs further explanation?

I happened to notice this chance for performance improvement.

In the `(root, result) => {}` function, `fs.readFileSync()` is called many times for each scanned file.
However, `functionsList` is immutable, so we can call it only once.

Benchmark:

```sh-session
$ cat tmp/stylelint.config.mjs
export default {
	rules: { 'function-no-unknown': true },
};

$ cat tmp/test.css
a {
  width: cal(1 + 2);
}

$ for i in $(seq 1 1000); do cp tmp/test.css "tmp/test-${i}.css"; done

$ time bin/stylelint.mjs -c tmp/stylelint.config.mjs "tmp/*.css" 2>/dev/null
```

On the main branch:

```
0.92s user 0.32s system 133% cpu 0.928 total
```

On this branch:

```
0.88s user 0.29s system 132% cpu 0.878 total
```