@Mouvedia 2024-06-24
I don't think it's worth it, but I thought I would mention it nonetheless.
If you were to change the `tsconfig.json` like so

```
{
	"compilerOptions": {
		"module": "NodeNext",
		"moduleResolution": "NodeNext",
		"resolveJsonModule": true,
		…
```

then you could do

```js
import foo from 'css-functions-list/index.json' assert { type: 'json' };
```