@ybiquitous 2024-06-24
Unfortunately, it seems we cannot use [Import attributes](https://nodejs.org/api/esm.html#import-attributes) yet because:

- Node.js 18 doesn't support the feature.
- Node.js 20 supports it, but it's experimental.

See:

```sh-session
$ cat test-import-attributes.mjs
import functionsList from 'css-functions-list/index.json' with { type: 'json' };

console.log(functionsList[0]);

$ nvm exec 20 node test-import-attributes.mjs
Running node v20.14.0 (npm v10.8.1)
abs
(node:55164) ExperimentalWarning: Importing JSON modules is an experimental feature and might change at any time
(Use `node --trace-warnings ...` to show where the warning was created)

$ nvm exec 18 node test-import-attributes.mjs
Running node v18.12.0 (npm v8.19.2)
file:///Users/masafumi.koba/git/stylelint/stylelint/test-import-attributes.mjs:1
import functionsList from 'css-functions-list/index.json' with { type: 'json' };
                                                          ^^^^

SyntaxError: Unexpected token 'with'
    at ESMLoader.moduleStrategy (node:internal/modules/esm/translators:119:18)
    at ESMLoader.moduleProvider (node:internal/modules/esm/loader:468:14)
    at async link (node:internal/modules/esm/module_job:67:21)

Node.js v18.12.0
```