@Mouvedia 2024-06-24
That's why I was using `assert` and not `with` but it's not necessary in our case.
i.e. just `"resolveJsonModule": true` with `import foo from 'css-functions-list/index.json';` is sufficient