@Mouvedia 2024-06-24
There's also

```js
import { createRequire } from 'node:module';
const require = createRequire(import.meta.url);

const foo = require('css-functions-list/index.json');
```