@ybiquitous 2024-06-24
@Mouvedia Your suggestion sounds better since it's the standard way to import a JSON file. Please take a look at 4fd66f901 again. 🙏🏼