@ybiquitous 2024-06-24
> Have you seen what rollup is generating?
> That's surprisingly inefficient.

Yes, this is a known thing. Rollup transforms `createRequire` to such.